11456 - Trainsorting
LIS and LDS from back to front.

It means that LIS[i] is a size of a longest increasing subsequence which starts at index i, and its end point is the last element.

Originally, LIS[i] is a size of a longest increasing subsequence which starts at 0, and its end point is index i.